DJANGO REST NOTES/CALL API FOR CRUD OPERATION/requirements.txt
asgiref==3.5.1
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.1.0
